+++ src/test/java/step3/racingCar/GameTest.java
@@ -3,6 +3,8 @@ package step3.racingCar;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.MockedStatic;
+import step3.racingCar.domain.Cars;
+import step3.racingCar.ui.InputView;
 
 import java.util.ArrayList;
